TransactionID types
