+++ packages/ui/src/2_molecules/TransactionID/TransactionID.tsx
@@ -5,38 +5,9 @@ import classNames from 'classnames';
 import { Icon } from '../../1_atoms';
 import { COPY, NEW_TAB } from '../../1_atoms/Icon/iconNames';
 import { Tooltip } from '../Tooltip/Tooltip';
-import { TooltipPlacement } from '../Tooltip/Tooltip.types';
+import { TooltipPlacement, TooltipTrigger } from '../Tooltip/Tooltip.types';
 import styles from './TransactionID.module.css';
-
-export const chains = {
-  mainnet: 30,
-  testnet: 31,
-};
-
-export const currentChainId = chains.mainnet;
-
-export enum ChainId {
-  ETH_MAINNET = 1,
-  ETH_TESTNET = 3, // Ropsten
-  ETH_RINKEBY = 4,
-  RSK_MAINNET = 30,
-  RSK_TESTNET = 31,
-  BSC_MAINNET = 56,
-  BSC_TESTNET = 97,
-  MATIC_MAINNET = 137, // Polygon
-  MATIC_TESTNET = 80001,
-}
-
-export const blockExplorers = {
-  1: 'https://etherscan.io',
-  3: 'https://ropsten.etherscan.io',
-  30: 'https://explorer.rsk.co',
-  31: 'https://explorer.testnet.rsk.co',
-  btc_30: 'https://live.blockcypher.com/btc',
-  btc_31: 'https://live.blockcypher.com/btc-testnet',
-  56: 'https://bscscan.com',
-  97: 'https://testnet.bscscan.com',
-};
+import { blockExplorers, ChainId, currentChainId } from './TransactionID.types';
 
 interface TransactionIDProps {
   value: string;
@@ -45,7 +16,7 @@ interface TransactionIDProps {
   className?: string;
   hideTooltip?: boolean;
   chainId?: number | ChainId;
-  dataActionId?: string;
+  dataLayoutId?: string;
   isNativeBtc?: boolean;
   isAddress?: boolean;
 }
@@ -56,7 +27,7 @@ export const TransactionID: React.FC<TransactionIDProps> = ({
   className,
   hideTooltip,
   chainId,
-  dataActionId,
+  dataLayoutId,
   isNativeBtc,
   isAddress,
   ...props
@@ -104,10 +75,12 @@ export const TransactionID: React.FC<TransactionIDProps> = ({
           </a>
         </span>
       }
+      trigger={TooltipTrigger.click}
       className={classNames(styles.link, className)}
       tooltipClassName={styles.tooltip}
-      dataActionId={dataActionId}
+      dataLayoutId={dataLayoutId}
       placement={TooltipPlacement.BOTTOM_END}
+      disabled={hideTooltip}
     >
       <div>{value}</div>
     </Tooltip>
